+++ android/app/src/main/kotlin/com/temanlabs/temanku/MainActivity.kt
@@ -1,0 +1,5 @@
+package com.temanlabs.temanku
+
+import io.flutter.embedding.android.FlutterActivity
+
+class MainActivity : FlutterActivity()